requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
greenlet==3.0.0

# Azure AI (supports both OpenAI and AI Inference)
openai==1.3.7
azure-ai-inference==1.0.0b4
tiktoken==0.5.2

# WebSocket support
websockets==12.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# Template engine
jinja2==3.1.2

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
structlog==23.2.0

# Image processing
Pillow==10.1.0
pillow-heif==0.13.0
python-magic==0.4.27

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
beautifulsoup4~=4.13.4